+++ Cargo.toml
@@ -35,7 +35,3 @@ features = [
   'RequestCache',
 ]
 
-[dependencies.ical]
-version = "0.8.*"
-default-features = false
-features = ["ical", "vcard"]
