+++ index.html
@@ -1,6 +1,5 @@
 <!DOCTYPE html>
 <link data-trunk rel="copy-file" href="Bahn.ttf"/>
-<link data-trunk rel="copy-file" href="zeiten.json"/>
 <style>
 @import url('https://fonts.cdnfonts.com/css/5by7');
 @-moz-keyframes my-animation {
