--- zeiten.json
@@ -1,10 +1,0 @@
-{
-    "times": [
-        {
-            "title": "WISe 23/24",
-            "start": "2023-10-01T00:00:00Z",
-            "end": "2024-02-02T00:00:00Z"
-        }
-    ],
-}   
-
